requirements.txt
biopython
torch
numpy
scikit-learn